change Number of Depriciation from float to integer

+++ asset_account_hcv_extension/asset.py
@@ -42,8 +42,8 @@ class account_asset_category(osv.osv):
 
 	_columns = {
 		'method_linear_factor': fields.float('Linear Factor'),
-		'method_number_readonly': fields.related('method_number', type='float', string='Number of Depriciations', store=True, readonly=True),
-		'method_number': fields.float('Number of Depreciations', help="The number of depreciations needed to depreciate your asset"),
+		'method_number_readonly': fields.related('method_number', type='integer', string='Number of Depriciations', store=True, readonly=True),
+		'method_number': fields.integer('Number of Depreciations', help="The number of depreciations needed to depreciate your asset"),
 	}
 
 	def onchange_method(self, cr, uid, ids, method, factor, context=None):
